feat: move to my reading components
